Introduce non-streaming APIs
HighwayHash is a fast enough hash function that the performance penalty
of leaving the land of JS is tangible.

To give an example, the snippet has 3 N-API calls:

```ts
function hashData64(data, hash, key) {
    const hasher = hash.create(key);
    hasher.append(data);
    return hasher.finalize64();
}
```

If we replace it with a single N-API call that initializes, hashes, and
finalizes in a single step there can be major performance benefits:

- Hashing data with a length of 1 doubled in throughput (0.17 MB/s to
0.38 MB/s)
- Even at larger input sizes (~1 MB) there can be a throughput gap of
10-15% between.

So when performance is key, one should design their API to allow users
to call methods that reduce the number of N-API calls.