Enable cross compiling
This introduces the ability to cross compile the highwayhasher package
to other architectures than what CI natively supports. This is done via
Cross. This is in stark contrast to how other napi-rs packages are cross
compiled. Truthfully I like the cross approach better. It's what I know
and it seems more straightforward.

Closes #1