Use simple JS in browser tests to fix bundle errors
For the past month the browser tests have been failing with the
following error:

```
Can't resolve 'core-js/modules/web.dom.iterable.js'
```

I'm not sure what precipated this change as CI all of a sudden starting
failing without any code changes.

Searching on the web was not fruitful as the peddled solutions involved
installing `core-js` (2 or 3) and neither worked.

What is interesting, is that using a plain for-loop over iterators and
desctructuring fixed the issue.